* enable edit content

+++ src/main/java/org/knowledgeroot/app/content/ContentDao.java
@@ -16,7 +16,7 @@ public interface ContentDao {
      * @param id
      * @return
      */
-    Content findById(long id);
+    Content findById(Integer id);
 
     /**
      * check if content exists
